fix(location-service): handle incase id doesn't in delete user (55-0654)

+++ location-service/src/main/java/com/team01/uber/location/service/LocationService.java
@@ -44,6 +44,9 @@ public class LocationService {
     }
 
     public void delete(Long id) {
+        if (!locationRepository.existsById(id)) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Error 404");
+        }
         locationRepository.deleteById(id);
     }
 }
